// src/app/interfaces/elemento.ts
export interface Elemento {
    Cantidad: number;      // Cantidad del elemento
    Estado: boolean;       // Estado del elemento (activo/inactivo)
    IdSitio: number;       // ID del sitio asociado
    completado: boolean;   // Estado de completado del elemento
    nombre: string;        // Nombre del elemento
    id:number;
}
  